requirements.txt
astroid==2.4.1
autopep8==1.5.2
certifi==2018.8.24
chardet==3.0.4
defusedxml==0.5.0
dj-database-url==0.5.0
Django==2.2
django-allauth==0.37.1
django-cors-headers==2.4.0
django-heroku==0.3.1
django-rest-auth==0.9.5
djangorestframework==3.8.2
gunicorn==19.7.1
idna==2.7
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
oauthlib==2.1.0
psycopg2==2.8.5
pycodestyle==2.5.0
pylint==2.5.2
python3-openid==3.1.0
pytz==2018.5
requests==2.21.0
requests-oauthlib==1.0.0
six==1.14.0
sqlparse==0.3.0
toml==0.10.0
typed-ast==1.4.1
urllib3==1.24.2
whitenoise==3.3.1
wrapt==1.12.1
